requirements.txt
streamlit
numpy
pillow
streamlit-image-coordinates

